added fields to the activities. is_enabled and user_id

+++ app/Http/Controllers/ActivityController.php
@@ -16,6 +16,7 @@ class ActivityController extends Controller
     {
         $activities = Activity::query()
             ->notStarted()
+            ->active()
             ->with('users', 'trainers')
             ->orderBy('start_time', 'asc')
             ->get();
@@ -25,14 +26,14 @@ class ActivityController extends Controller
 
     public function my()
     {
-        $activities = auth()->user()->activities()->notStarted()->get();
+        $activities = auth()->user()->activities()->notStarted()->active()->get();
 
         return view('site.activity.my', compact('activities'));
     }
 
     public function myArchive()
     {
-        $activities = auth()->user()->activities()->old()->get();
+        $activities = auth()->user()->activities()->old()->active()->get();
 
         return view('site.activity.my-archive', compact('activities'));
     }
@@ -40,6 +41,7 @@ class ActivityController extends Controller
     public function show($id)
     {
         $activity = Activity::query()
+            ->active()
             ->with('users', 'trainers')
             ->where('id', $id)->first();
 
@@ -49,14 +51,24 @@ class ActivityController extends Controller
     public function join($id)
     {
         $activity = Activity::query()->where('id', $id)->first();
-        $activity->users()->attach(auth()->id());
-        return back()->with('success', 'Успишно приєднано!');
+
+        if($activity) {
+            $activity->users()->attach(auth()->id());
+            return back()->with('success', 'Успишно приєднано!');
+        }
+
+        return back()->with('error', 'Немає такого заняття');
     }
 
     public function cancelJoin($id)
     {
         $activity = Activity::query()->where('id', $id)->first();
-        $activity->users()->detach(auth()->id());
-        return back()->with('success', 'Успішно скасовано!');
+
+        if($activity) {
+            $activity->users()->detach(auth()->id());
+            return back()->with('success', 'Успишно приєднано!');
+        }
+
+        return back()->with('error', 'Немає такого заняття');
     }
 }
